Map tweaks
- Fixed issue relating to polygon misalignment on zoom.

It's a bit of a hacky workaround, but it works pretty much flawlessly (for now).

- Changed polygon behaviour.

Added a toggle button for polygons.

Instead of being hidden on the mouse enter, the polygons will stay present unless the map is actively being moved, allowing people to move the map and still see the boundaries of their ticket.